qmanager: consider remaining jobs schedulable after ENODEV failure
Problem: the current qmanager implementation of first-come first-
serve queue policy batches multiple jobs together when trying to
make matches, evaluating them one at a time. When an unsatisfiable
job is submitted, ENODEV is returned. When many are submitted,
the requests are batched together, but only one ENODEV is received
before the scheduler goes to sleep, considering the rest of the
batch "unschedulable."

Solution: make sure if ENODEV is received, it triggers the
conditional path which results in set_schdulability(true).

Co-authored-by: Tom Scogland <scogland1@llnl.gov>